+++ CHANGELOG.md
@@ -4,6 +4,7 @@ NextJS Starter is a boilerplate project that serves as a starting point for all 
 
 ## Upcoming Release: v1.0.0
 
+- [#8](https://github.com/jtcarlos/nextjs-starter/issues/8) - app(feat): add page footer component
 - [#7](https://github.com/jtcarlos/nextjs-starter/issues/7) - app(feat): add page navigation component
 - [#6](https://github.com/jtcarlos/nextjs-starter/issues/6) - app(feat): add page layout component
 - [#5](https://github.com/jtcarlos/nextjs-starter/issues/5) - app(feat): add helmet component